move authentication to services
